requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
gspread
oauth2client
python-dotenv
python-dateutil
